cover: refact -> move helper function to shared
